Upgrade Gradle to v9.0 and fix build scripts issues
This commit upgrades the Gradle version to 9.0.0.
Additionally, it refactors the build scripts in `magisk-loader` and `apache` modules to use the `tasks.register` API for task creation.
The `androidResources` build feature is now explicitly disabled using the new syntax in `dex2oat` and `core` modules.
To fix the gradle-lint issues, the `androidx.startup` dependency is added to the `app` module.